ref/moved SSG apis to services
