+++ src/pages/article.tsx
@@ -1,10 +1,8 @@
-import axios, { AxiosResponse } from 'axios';
 import { Box } from '@chakra-ui/react';
-import openGraphScraper from 'open-graph-scraper';
 import { Title } from 'ui/title';
 import { ArticleLayout } from 'ui/article/articleLayout';
-import type { OGPDataType, GetArticlesResponse } from 'utils/types';
-
+import { getArticleAPI } from 'services/get-article-api';
+import { OGPDataType } from 'services/get-ogp';
 
 interface Props {
   articles: Array<OGPDataType>;
@@ -19,44 +17,14 @@ const ArticleContainer = ({ articles }: Props): JSX.Element => {
   );
 };
 
-export const getStaticProps = async () => {
-  const articles = await axios
-    .get(process.env.NEXT_PUBLIC_ARTICLE_URL, {
-      headers: { 'X-API-KEY': process.env.NEXT_PUBLIC_X_API_KEY },
-    })
-    .then(({ data }: AxiosResponse<GetArticlesResponse>) => {
-      return Promise.all(data.contents.map((content) => getOGPData(content)));
-    })
-    .catch((err) => null);
+export const getStaticProps = async (): Promise<{ props: Props }> => {
+  const articles = await getArticleAPI();
 
   return {
     props: {
       articles,
     },
   };
-};
-
-const getOGPData = async (article) => {
-  const data = await openGraphScraper({
-    url: article.url,
-    timeout: 10000,
-    onlyGetOpenGraphInfo: true,
-  });
-
-  if (!data.result.success || data.error ) {
-    return Promise.resolve({ 
-      url: article.url,
-      title: "検索中",
-      image: "/not_found.png",
-    });
-  }
-
-
-  return Promise.resolve({
-    url: article.url,
-    title: data.result.ogTitle,
-    image: data.result.ogImage,
-  });
 };
 
 export default ArticleContainer;
