+++ src/services/client.ts
@@ -1,0 +1,15 @@
+import axios, { Method } from 'axios';
+
+interface RequestConfig {
+  method: Method;
+  url: string;
+  data?: any;
+}
+
+export const client = (config: RequestConfig) =>
+  axios({
+    ...config,
+    headers: { 'X-API-KEY': process.env.NEXT_PUBLIC_X_API_KEY },
+  });
+
+export type Client = ReturnType<typeof client>;
